feat: move tasks state to scrollList state
