+++ src/components/DayColumn/DayColumn.types.ts
@@ -2,4 +2,6 @@ import { DateTime } from "luxon";
 
 export interface DayColumnProps {
   monday: DateTime;
+  addTask: (e:any) => void;
+  task: string[];
 }
